the walker - V0.1
still lacking much of the texturing and shading. I wanted to draw inspiration from the surrealism aesthetics for this artwork. the rule-breaking creativity of artists like Magritte and Dali quite align with what Fedora represents for me.
also, the representation of the community is completely missing, which is the key part of this illustration... but it's also the bit that would take me the longest to pull off, therefore I'd get around to doing it only if this direction is relevant at all :)


check the inspo board and my initial sketches, how the community will be manifested in the middle of the screen: